Added "Game Over" and Sequence Increase Functionality
When the player clicks the wrong button, a pop-up will appear stating
that there has been a "game over". This will reset the game back to
initial conditions.

If the player correctly clicks the buttons in the correct sequence, the
progress bar will reach 100%, there will be a short pause, then a longer
sequence will play, indicating the next level.